pyproject.toml
[tool.poetry]
name = "jobpulse"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "jobhunter"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
beautifulsoup4 = "^4.11.2"
nltk = "^3.8.1"
coverage = "^5.5"
mypy = "^1.6.1"
pandas = "^2.0.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
scikit_learn = "^1.2.1"
streamlit = "^1.30.0"
tqdm = "^4.65.0"
gensim = "^4.3.2"
PyPDF2 = "^1.27.0"
scipy = "^1.11.2"
pdfplumber = "^0.11.6"
sentence-transformers = "^3.0.0"
python-jobspy = "^1.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"

[tool.poetry.scripts]
jobhunter = "jobhunter.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
